requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bleach==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.4
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filebrowser==4.0.2
django-grappelli==3.0.5
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-rest-multiple-models==2.1.3
django-solo==2.0.0
django-summernote==0.8.20.0
django-tinymce==3.5.0
django-unfold==0.5.0
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.12.2
gunicorn==20.1.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.0
# Pillow==9.4.0
# psycopg2-binary==2.9.5
pytz==2022.7
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
# safetensors==0.4.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
sympy==1.12
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.3.0
